+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-* Create NoArgFormatStringAnalyzer, reports printf-style functions applied to a format string without format specifiers.
+* Create NoArgFormatStringAnalyzer, reports printf-style functions (including `fprintf`, `fprintfn` and `bprintf`) applied to a format string without format specifiers.
 
 ## [0.18.0] - 2026-09-08